Refactor TagsCell changes to use Concern
